Add cytoband to the svg export also
